General Refactoring
Starting to move locations to end of argument lists to make easier transitions to productions, using AGDcl over AGDcls, declaring too many locals for the transform dcl forward instead of recursively crashing. Disabling some things to attempt to get the output to work-- running into the obvious problem of attempting to check the environment for defined elements as we are defining new elements, recursively checking this dcl.