Change fst and snd to annotations.
This is part of a re-bootstrap with its parent commit. This makes the
library changes necesssary to get a compiler that builds again. However,
it needs updates to the runtime and tests still.

In the bootstrap, this means we need to build this commit, then update
the runtime before using it.