A perhaps slightly hacky fix for "Ambiguous reference to type 'fully:Qualified'" errors.
In a situation like:

    grammar foo;
    nt Foo;

    grammar bar;
    exports foo;

    grammar baz;
    import foo;
    import bar;
    type Baz = foo:Foo;

Despite fully qualifying foo:Foo, we still get an ambiguous-reference
error, since we could be getting foo:Foo from the import of foo or from
the import of bar, since bar exports foo.

We fix this by sorting and uniq-ing the results of environment queries,
using the full name as a key. It should never (right? right??) be the
case that:

1. there are multiple items with the same fullName in the environment
2. the only place this is checked is the lookup of that fullName

so this ought to be a sound fix.

top.dcls should always have exactly one element in a non-error case, so
the sort and uniq calls should be cheap. In theory, only the uniq is
necessary, but I guess this could improve error message quality in some
other case too. A <=1-element list is the base case of the current
implementation of sort, though, so yeah, cheap.